Guardado en base y mostrado de la actividad de sus juegos

+++ app/src/main/java/ami/coach/game/gamingcoach/RegistroActivity.java
@@ -36,12 +36,12 @@ import javax.xml.parsers.DocumentBuilderFactory;
 /**
  * Created by ces_m on 12/22/2015.
  */
-public class RegistroActivity extends AppCompatActivity {
+public class RegistroActivity extends AppCompatActivity{
 
     ImageView boton_log;
     EditText id;
     Context ctx;
-    Activity activity;
+    RegistroActivity activity;
     ProgressDialog consultando;
     public static final String MyPREFERENCES = "logPreferences" ;
     private SharedPreferences sharedpreferences;
@@ -123,6 +123,19 @@ public class RegistroActivity extends AppCompatActivity {
         consultando.show();
     }
 
+    public void consultarJuegos(final AsyncTask hilo){
+        if(consultando!=null)consultando.dismiss();
+        consultando = new ProgressDialog(activity);
+        consultando.setTitle("Consultando lista de juegos...");
+        consultando.setMessage("Espere mientras se consulta los juegos...");
+        consultando.setCanceledOnTouchOutside(false);
+        consultando.setCancelable(true);
+        consultando.setButton(DialogInterface.BUTTON_NEGATIVE, "Cancelar", new DialogInterface.OnClickListener() {
+            @Override public void onClick(DialogInterface dialog, int which){dialog.dismiss();hilo.cancel(true);}
+        });
+        consultando.show();
+    }
+
 
     public class GetInfoXml extends AsyncTask<String, Void, Object[]> {
 
@@ -164,26 +177,31 @@ public class RegistroActivity extends AppCompatActivity {
                 list = doc.getElementsByTagName("avatarFull");
                 node = list.item(0);
 
-                //URL url = new URL(node.getTextContent());
-                URL url = new URL("https://www.google.com.ec/logos/doodles/2015/holidays-2015-day-1-6575248619077632-hp.jpg");
+                URL url = new URL(node.getTextContent());
+                //URL url = new URL("https://www.google.com.ec/logos/doodles/2015/holidays-2015-day-1-6575248619077632-hp.jpg");
                 InputStream input = url.openConnection().getInputStream();
 
 
                 File path = new File(Environment.getExternalStorageDirectory().getAbsolutePath() + "/GamingCoach/avatars");
                 File mediaFile = new File(path, customUrl+".jpg");
-                path.mkdirs();
+                if(!mediaFile.exists()){
 
-                OutputStream os = new FileOutputStream(mediaFile);
-                byte data[] = new byte[4096];
-                int count;
-                while ((count = input.read(data)) != -1) {
-                    os.write(data, 0, count);
+                    path.mkdirs();
+
+                    OutputStream os = new FileOutputStream(mediaFile);
+                    byte data[] = new byte[4096];
+                    int count;
+                    while ((count = input.read(data)) != -1) {
+                        os.write(data, 0, count);
+                    }
+
+                    if (os != null)
+                        os.close();
+                    if (input != null)
+                        input.close();
+
                 }
 
-                if (os != null)
-                    os.close();
-                if (input != null)
-                    input.close();
 
 
                 result[3]=mediaFile.getAbsoluteFile();
@@ -245,10 +263,6 @@ public class RegistroActivity extends AppCompatActivity {
                 return;
             }
 
-            Intent intent = new Intent(activity, MainActivity.class);
-            startActivity(intent);
-
-
             Date date = new Date(System.currentTimeMillis());
             long millis = date.getTime();
 
@@ -258,11 +272,17 @@ public class RegistroActivity extends AppCompatActivity {
             editor.putString(Prefs.OnlineState.name(), result[1].toString());
             editor.putString(Prefs.StateMessage.name(), result[2].toString());
             editor.putString(Prefs.Avatar.name(), result[3].toString());
-            editor.putLong(Prefs.updated.name(),millis);
+            editor.putLong(Prefs.updated.name(), millis);
             editor.putString(Prefs.CustomUrl.name(), customUrl);
             editor.putBoolean(Prefs.UserLog.name(), true);
             editor.commit();
-            activity.finish();
+
+            GetJuegosXml getJuegosXml = new GetJuegosXml();
+            getJuegosXml.activity=activity;
+            consultarJuegos(getJuegosXml);
+            getJuegosXml.execute(customUrl, result[5].toString());
+
+            //activity.finish();
 
         }
 
@@ -271,5 +291,10 @@ public class RegistroActivity extends AppCompatActivity {
         }
     }
 
-
+    public void procesoFinal() {
+        //if(consultando!=null)consultando.dismiss();
+        Intent intent = new Intent(activity, MainActivity.class);
+        startActivity(intent);
+        activity.finish();
+    }
 }
